Added argparser option
Before, we needed two different versions for the VLBI monitor and VLBI Field System. This update allows the user to pass the localfolder where to copy and read the xml cals files (or not pass an argument to read directly from the telescope data folder). Also the -v option allows to select verbosity or not.