Added mandatory option to choose output format of ANTAB table depending if GMVA or EHT session is being processed.
This is required because (at Sep2018), GMVA records 2 IFs, but EHT records 4 IFs. Thus, both the header  and the number of columns of the ANTAB table is different for each type of session.

This is now automatically taken care by the code. The user must provide the last argument as {GMVA or EHT}.

Nothe that, for minimal code modification, it is best that even in a GMVA session, 4 IFs are connected (e.g. E090+E230), so the FS logs get 4 lines per calibration info. The code currently expects 4 lines per cal. scan.